MissileLauncher, neue Level, Turret unlock
MissileLauncherUpgrade schiesst zweimal, 3neue Level noch nicht designed, Tuerme werden von Level zu Level freigeschaltet :apple:

+++ Game Design/Assets/Scripts/Bullet.cs
@@ -8,7 +8,7 @@ public class Bullet : MonoBehaviour
     public GameObject impactEffect;
 
     public float speed = 60f;
-    public int damage = 20;
+    public float damage = 20;
     public float explosionRadius = 0f;
     public AudioClip missileLauncherExplosionSfx;
